requirements.txt
numpy~=1.23.5
pandas~=1.5.3
easydict~=1.10
plotly~=5.14.1
seaborn~=0.13.2
python-dotenv~=1.0.1
opencv-python~=4.9.0.80
pillow~=9.5.0
scikit-image~=0.22.0
